+++ src/client/components/Home.js
@@ -2,22 +2,21 @@ import React, { useState } from "react"
 import MealItem from "./MealItem"
 import FetchAPI from "./FetchAPI"
 
-export default function Meal({ id }) {
+export default function Home() {
   const [loading, setLoading] = useState(true)
   const [data, setData] = useState("")
   return (
     <div>
-      <h1>Meal</h1>
-      <h2>Render the meal</h2>
+      <h1>Home</h1>
+      <h2>Render all the meals</h2>
       {/* When the data is loading, show "loading..." */}
       {loading ?
         <>
           "loading..."
-          <FetchAPI setData={setData} setLoading={setLoading} API={"http://localhost:5000/api/meals/" + id} />
-        </>
-        :
+          <FetchAPI setData={setData} setLoading={setLoading} API={"http://localhost:5000/api/meals"} />
+        </> :
         <>
-          {/* When the fetch result is empty (no meals), show "No results..." */}
+          {/* When the data result is empty (no meals), show "No results..." */}
           {data.length === 0 ? "No results..." :
             <>
               <ul>
