+++ src/client/components/MealItem.js
