--- src/client/components/TestComponent/TestComponent.js
@@ -1,13 +1,0 @@
-/* eslint-disable camelcase */
-/* eslint-disable @typescript-eslint/camelcase */
-import React from "react";
-import PropTypes from "prop-types";
-import "./testComponentStyle.css";
-
-export default function TestComponent() {
-  return (
-    <section className="test-component">
-      <p>in a component</p>
-    </section>
-  );
-}
