--- src/client/components/TestComponent/testComponentStyle.css
@@ -1,3 +1,0 @@
-section.test-component {
-  color: green;
-}
